+++ docs/changelog.rst
@@ -5,6 +5,10 @@ Changelog
 =========
 
 
+- :release:`10.3.0 <19th September 2023>`
+- :feature:`195` Add `log_format` to `pydis_core.utils.logging` to allow for standardised logging across all services using pydis_core.
+
+
 - :release:`10.2.0 <28th August 2023>`
 - :support:`192` Bump Discord.py to :literal-url:`2.3.2 <https://github.com/Rapptz/discord.py/releases/tag/v2.3.2>`.